completed first iteration of ingestion flow

+++ src/index.js
@@ -4,12 +4,8 @@ const fs = require('fs');
 const path = require('path');
 const mime = require('mime-types');
 const argv = require('minimist')(process.argv.slice(2));
-const {ingestXlsx} = require('./ingest_helpers');
+const {ingestReport} = require('./ingest_helpers');
 const {throwError, deleteReport} = require('./utils');
-
-const directory = './ingested_sales_reports';
-const file = 'report.json';
-const reportExists = fs.existsSync(path.join(directory, 'report.json'));
 
 const acceptedMimeTypes = [
   'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
@@ -17,7 +13,7 @@ const acceptedMimeTypes = [
 ];
 
 if (argv.ingest) {
-  // this allows a filepath or filename to be passed in
+  // this allows a filepath or just the filename to be passed in
   const report = argv.ingest.includes('/') ? argv.ingest : path.join(__dirname, argv.ingest);
   const mimeType = mime.contentType(path.extname(report));
 
@@ -25,16 +21,19 @@ if (argv.ingest) {
     throwError({message: 'Error: Only .xlsx & .txt can be ingested'});
   }
 
+  const directory = './ingested_sales_reports';
+  const file = path.basename(report).split('.')[0];
+  const reportExists = fs.existsSync(path.join(directory, `${file}.json`));
+
   try {
-    if (path.extname(report) === '.xlsx') {
-      ingestXlsx(report, reportExists, directory, file);
-    } else {
-      console.log('ingest .txt');
-    }
+    ingestReport(report, reportExists, directory, file);
   } catch (error) {
-    if (reportExists) {
+    if (reportExists && error.code !== 'ENOENT') {
       deleteReport(directory, file);
     }
+
     throwError({message: error});
   }
+} else if (argv.summary) {
+  console.log
 }
